+++ src/components/MainPage/MainPage.tsx
@@ -1,8 +1,10 @@
 import { Inter } from 'next/font/google';
 import { useRouter } from 'next/router';
+import { FC } from 'react';
 import { useDispatch } from 'react-redux';
 
 import { updatePage } from '@/app/store/actions/pageSlice.ts';
+import { Artwork, IResponseDetails } from '@/types/types.ts';
 
 import styles from './MainPage.module.css';
 import { ErrorButton } from '../ErrorButton/ErrorButton.tsx';
@@ -11,7 +13,12 @@ import Search from '../SearchSection/search.tsx';
 
 const inter = Inter({ subsets: ['latin'] });
 
-const MainPage = () => {
+interface MainPageProps {
+  arrArtworks: Array<Artwork>;
+  details?: IResponseDetails;
+}
+
+const MainPage: FC<MainPageProps> = ({ arrArtworks, details }) => {
   const dispatch = useDispatch();
   const router = useRouter();
 
@@ -24,7 +31,7 @@ const MainPage = () => {
         }}
       />
       <main className={styles.main}>
-        <SearchResultsSection />
+        <SearchResultsSection arrArtworks={arrArtworks} details={details} />
       </main>
       <ErrorButton />
     </div>
