+++ src/components/SearchResultsSection/SearchResultsSection.tsx
@@ -1,23 +1,26 @@
 import { useRouter } from 'next/router';
-import { useSelector } from 'react-redux';
+import { FC } from 'react';
 
-import { RootState } from '@/app/store/store.ts';
 import Details from '@/components/details/Details.tsx';
+import { Artwork } from '@/types/types.ts';
 
 import styles from './SearchResultsSection.module.css';
-import ListOfCardWithPagination from '../ListOfCardWithPagination/ListOfCardWithPagination.tsx';
+import ListOfCard from '../ListOfCardWithPagination/ListOfCardWithPagination.tsx';
 
-const SearchResultsSection = () => {
-  const arrValue = useSelector(
-    (state: RootState) => state.arrArtworks.arrArtworks
-  );
+interface SearchResultsSectionProps {
+  arrArtworks: Array<Artwork>;
+}
+
+const SearchResultsSection: FC<SearchResultsSectionProps> = ({
+  arrArtworks,
+}) => {
   const router = useRouter();
   const { query } = router;
 
   return (
     <div className={styles.section}>
-      <ListOfCardWithPagination />
-      {arrValue.length > 0 ? (
+      <ListOfCard arrArtworks={arrArtworks} />
+      {arrArtworks.length > 0 ? (
         <div className={styles.detailsWrapper}>
           {query.details && <Details />}
         </div>
