--- src/components/listOfCard/listOfCard.module.css
@@ -1,48 +1,0 @@
-.cards {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 20px;
-  justify-content: center;
-  align-items: start;
-}
-
-.cardsOpenDetails {
-  display: grid;
-  gap: 20px;
-  justify-content: center;
-  align-items: start;
-  grid-template-columns: repeat(2, 1fr);
-}
-
-.card {
-  background: white;
-  display: flex;
-  flex-direction: column;
-  width: 250px;
-  height: 100%;
-  text-align: center;
-  justify-content: space-between;
-  box-shadow:
-    0 19px 38px rgba(0, 0, 0, 0.3),
-    0 15px 12px rgba(0, 0, 0, 0.3);
-  transition: 1s ease;
-}
-
-.card:hover {
-  cursor: pointer;
-  transform: scale(1.05);
-}
-
-.card img {
-  object-fit: contain;
-}
-
-.cardTitle {
-  padding: 0 5px;
-}
-
-.message {
-  color: white;
-  text-align: center;
-  font-size: 24px;
-}
